sync-github-releases: inline the target-commitish decision into apply
target-commitish.ts split a nine-line guard cascade into an impure
gather (resolveTargetCommitish) and a pure decide (decideTargetCommitish),
threaded through a five-field param object and a deduceCommit thunk. None
of that machinery was inherent to the logic — it existed solely so the
decision could be unit-tested without touching git, at the cost of
contorting the code that's actually read.

Inline it back into a single resolveTargetCommitish() guard cascade and
fold it into apply-release-plan.ts, its only caller, deleting the module.
The two refuse-to-tag safety paths (latest release untagged; older release
with no deducible commit) still earn coverage — those silent, sticky
failures have no e2e net — so they move to a git-mocked applyReleasePlan
test instead of a pure-function one.
