+++ .github/workflows/sync-github-releases/sync/apply-release-plan.spec.ts
@@ -2,33 +2,38 @@ import { describe, expect, it, vi } from 'vitest'
 import { applyReleasePlan } from './apply-release-plan.ts'
 import type { ReleasePlan } from './release-plan.ts'
 import type { Release, ReleasesClient } from '../utils/github.ts'
+import { findReleaseCommit, gitTagExists } from '../utils/git.ts'
+
+// Creating a release resolves its target commit by shelling out to git; stub that so the create path
+// (and its two refuse-to-tag failures) can run without a real repo. update/delete don't touch git.
+vi.mock('../utils/git.ts', () => ({ gitTagExists: vi.fn(), findReleaseCommit: vi.fn() }))
+
+function createFakeClient(): { client: ReleasesClient; calls: string[] } {
+  const calls: string[] = []
+  const client: ReleasesClient = {
+    async list() {
+      return [] as Release[]
+    },
+    async create() {
+      calls.push('create')
+    },
+    async update(releaseId) {
+      calls.push(`update:${releaseId}`)
+    },
+    async delete(releaseId) {
+      calls.push(`delete:${releaseId}`)
+    },
+  }
+  return { client, calls }
+}
 
 describe('applyReleasePlan()', () => {
-  // No creates in the plan, so resolveTargetCommitish() (which shells out to git) isn't exercised
-  // here. The dry-run gate is shared by all three actions via applyWrite(), so update + delete prove it.
+  // update + delete only (no creates), so the git stub stays untouched. The dry-run gate is shared by
+  // all three actions via applyWrite(), so update + delete prove it.
   const plan: ReleasePlan = {
     releasesToCreate: [],
     releasesToUpdate: [{ release_id: 1, tag_name: 'v1.0.0', body: 'Fresh notes' }],
     releasesToDelete: [{ release_id: 2, tag_name: 'v0.9.0' }],
-  }
-
-  function createFakeClient(): { client: ReleasesClient; calls: string[] } {
-    const calls: string[] = []
-    const client: ReleasesClient = {
-      async list() {
-        return [] as Release[]
-      },
-      async create() {
-        calls.push('create')
-      },
-      async update(releaseId) {
-        calls.push(`update:${releaseId}`)
-      },
-      async delete(releaseId) {
-        calls.push(`delete:${releaseId}`)
-      },
-    }
-    return { client, calls }
   }
 
   it('performs each write and logs it', async () => {
@@ -54,5 +59,35 @@ describe('applyReleasePlan()', () => {
       '[dry-run] Would delete release v0.9.0',
     ])
     log.mockRestore()
+  })
+
+  // A create whose git tag is missing must never be tagged at the wrong commit. These two hard fails
+  // are the tool's safety guards, and the create path has no e2e net — so they earn a test even though
+  // the dry-run gate can't reach them (resolution happens before it, so a dry run still fails fast).
+  describe('refuses to create a release whose git tag is missing', () => {
+    const createPlan = (isLatest: boolean): ReleasePlan => ({
+      releasesToCreate: [{ tag_name: 'v1.0.0', body: 'Notes', version: '1.0.0', isLatest }],
+      releasesToUpdate: [],
+      releasesToDelete: [],
+    })
+
+    it('hard-fails the latest release (it must already be tagged)', async () => {
+      vi.mocked(gitTagExists).mockReturnValue(false)
+      const { client } = createFakeClient()
+
+      await expect(
+        applyReleasePlan({ plan: createPlan(true), client, defaultBranch: 'main', dryRun: false }),
+      ).rejects.toThrow(/latest release must already be tagged/)
+    })
+
+    it("hard-fails an older release whose commit can't be deduced", async () => {
+      vi.mocked(gitTagExists).mockReturnValue(false)
+      vi.mocked(findReleaseCommit).mockReturnValue(null)
+      const { client } = createFakeClient()
+
+      await expect(
+        applyReleasePlan({ plan: createPlan(false), client, defaultBranch: 'main', dryRun: false }),
+      ).rejects.toThrow(/couldn't be deduced/)
+    })
   })
 })
