+++ .github/workflows/sync-github-releases/sync/apply-release-plan.ts
@@ -1,7 +1,7 @@
 export { applyReleasePlan }
 
-import type { ReleasePlan } from './release-plan.ts'
-import { resolveTargetCommitish } from './target-commitish.ts'
+import type { ReleasePlan, ReleaseToCreate } from './release-plan.ts'
+import { findReleaseCommit, gitTagExists } from '../utils/git.ts'
 import type { ReleasesClient } from '../utils/github.ts'
 
 // Carry out a plan from getReleasePlan() against GitHub: create, then update, then delete. The
@@ -41,6 +41,29 @@ async function applyReleasePlan({
 
   for (const release of plan.releasesToDelete)
     await applyWrite(dryRun, 'delete', release.tag_name, () => client.delete(release.release_id))
+}
+
+// The commit a to-be-created release is tagged at (its GitHub `target_commitish`):
+//  - Tag already exists: GitHub uses it and ignores target_commitish — pass the branch as a no-op.
+//  - Tag missing on the latest release: hard fail. The just-released version must already be tagged;
+//    tagging it now would point at the default branch's HEAD (the wrong commit).
+//  - Tag missing on an older (backfilled) release: tag the commit deduced from the changelog history,
+//    or hard fail if it couldn't be deduced — never silently tag the wrong commit.
+function resolveTargetCommitish(release: ReleaseToCreate, defaultBranch: string): string {
+  const { tag_name: releaseTag, version, isLatest } = release
+  if (gitTagExists(releaseTag)) return defaultBranch
+  const refusal = `Refusing to create release ${releaseTag}: its git tag is missing`
+  if (isLatest) {
+    throw new Error(
+      `${refusal}. The latest release must already be tagged — creating it now would tag the wrong commit (the default branch's HEAD).`,
+    )
+  }
+  const commit = findReleaseCommit(version)
+  if (!commit) {
+    throw new Error(`${refusal} and its release commit couldn't be deduced from the changelog history.`)
+  }
+  console.warn(`Tag ${releaseTag} is missing — creating its release at deduced commit ${commit}`)
+  return commit
 }
 
 // Perform one write against GitHub — or, under --dry-run, just announce it. Centralizing the gate here
